theEye/requirements.txt
Django==3.1.7
jsonfield==3.1.0
pytz==2021.1
dj-config-url==0.1.1
dj-database-url==0.5.0
gunicorn==20.0.4
whitenoise==5.2.0
djangorestframework==3.12.2
requests==2.25.1
psycopg2-binary==2.8.6
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework-jwt==1.11.0
PyJWT==2.0.1